Keep the utils index as a pure barrel
